Added the rest of the necessary files
The challenge I had for this particular activity was trying to figure out memory. I know it wasn't part of the activity requirements, but I figured that I should try to conserve memory as a practice, so I tried becoming more aware of registers to use. The other challenge was just remembering how to use GDB and figuring out what I was looking at.